+++ infra/lib/analytics/stack.ts
@@ -4,7 +4,7 @@
 import * as cdk from "aws-cdk-lib";
 import { Duration, Expiration, RemovalPolicy, Stack } from "aws-cdk-lib";
 import * as appsync from "aws-cdk-lib/aws-appsync";
-import * as cognito from "aws-cdk-lib/aws-cognito";
+import * as cognito_identity from "@aws-cdk/aws-cognito-identitypool-alpha";
 import * as dynamodb from "aws-cdk-lib/aws-dynamodb";
 import * as iam from "aws-cdk-lib/aws-iam";
 import * as kinesis from "aws-cdk-lib/aws-kinesis";
@@ -67,56 +67,25 @@ class AnalyticsIntegrationTestStackEnvironment extends IntegrationTestStackEnvir
     // Create the Cognito Identity Pool with permission to put events
     // to the `pinpointApp`.
 
-    const identityPool = new cognito.CfnIdentityPool(this, "IdentityPool", {
+    const identityPool = new cognito_identity.IdentityPool(this, "IdentityPool", {
       allowUnauthenticatedIdentities: true,
     });
 
-    const unauthenticatedRole = new iam.Role(this, "UnauthenticatedRole", {
-      description: "Default role for anonymous users",
-      assumedBy: new iam.FederatedPrincipal(
-        "cognito-identity.amazonaws.com",
-        {
-          StringEquals: {
-            "cognito-identity.amazonaws.com:aud": identityPool.ref,
-          },
-          "ForAnyValue:StringLike": {
-            "cognito-identity.amazonaws.com:amr": "unauthenticated",
-          },
-        },
-        "sts:AssumeRoleWithWebIdentity"
-      ),
-      inlinePolicies: {
-        pinpoint: new iam.PolicyDocument({
-          statements: [
-            new iam.PolicyStatement({
-              effect: iam.Effect.ALLOW,
-              actions: [
-                "mobiletargeting:PutEvents",
-                "mobiletargeting:UpdateEndpoint",
-              ],
-              resources: [
-                // All subresources on the Pinpoint app. This is how Amplify
-                // does it even though it can probably be tweaked for the
-                // the two actions above:
-                // https://docs.aws.amazon.com/pinpoint/latest/developerguide/permissions-actions.html
-                `${pinpointApp.attrArn}*`,
-              ],
-            }),
-          ],
-        }),
-      },
+    const pinpointStatement = new iam.PolicyStatement({
+      effect: iam.Effect.ALLOW,
+      actions: [
+        "mobiletargeting:PutEvents",
+        "mobiletargeting:UpdateEndpoint",
+      ],
+      resources: [
+        // All subresources on the Pinpoint app. This is how Amplify
+        // does it even though it can probably be tweaked for the
+        // the two actions above:
+        // https://docs.aws.amazon.com/pinpoint/latest/developerguide/permissions-actions.html
+        `${pinpointApp.attrArn}*`,
+      ],
     });
-
-    new cognito.CfnIdentityPoolRoleAttachment(
-      this,
-      "IdentityPoolRoleAttachment",
-      {
-        identityPoolId: identityPool.ref,
-        roles: {
-          unauthenticated: unauthenticatedRole.roleArn,
-        },
-      }
-    );
+    identityPool.unauthenticatedRole.addToPrincipalPolicy(pinpointStatement);
 
     // Create a Kinesis Data Stream and configure Pinpoint to stream
     // events to it.
@@ -264,7 +233,7 @@ class AnalyticsIntegrationTestStackEnvironment extends IntegrationTestStackEnvir
       },
       authConfig: {
         identityPoolConfig: {
-          identityPoolId: identityPool.ref,
+          identityPoolId: identityPool.identityPoolId,
         },
       },
     };