+++ packages/aft/lib/src/commands/constraints_command.dart
@@ -1,11 +1,13 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:aft/aft.dart';
 import 'package:aft/src/options/glob_options.dart';
 import 'package:collection/collection.dart';
+import 'package:pub_api_client/pub_api_client.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:pubspec_parse/pubspec_parse.dart';
 import 'package:yaml_edit/yaml_edit.dart';
@@ -37,6 +39,7 @@ class ConstraintsCommand extends AmplifyCommand {
     addSubcommand(_ConstraintsSubcommand(_ConstraintsAction.check));
     addSubcommand(_ConstraintsSubcommand(_ConstraintsAction.apply));
     addSubcommand(_ConstraintsUpdateCommand());
+    addSubcommand(_ConstraintsPubVerifyCommand());
   }
 
   @override
@@ -332,5 +335,134 @@ class _ConstraintsUpdateCommand extends _ConstraintsSubcommand {
     if (hasUpdates && failedUpdates.isEmpty) {
       await _run(_ConstraintsAction.apply);
     }
+  }
+}
+
+class _ConstraintsPubVerifyCommand extends AmplifyCommand {
+  _ConstraintsPubVerifyCommand() {
+    argParser.addOption(
+      'count',
+      help: 'The number of pub packages to verify against',
+      defaultsTo: '100',
+    );
+  }
+  @override
+  String get name => 'pub-verify';
+
+  @override
+  String get description =>
+      'Verifies Amplify constraints against the top pub.dev packages';
+
+  final _pubClient = PubClient();
+
+  late final count = int.parse(argResults!['count'] as String);
+
+  @override
+  Future<void> run() async {
+    await super.run();
+
+    // Packages with version constraints so old, we consider them abandoned
+    // and don't bother running the constraint check for them.
+    const unofficiallyAbandonedPackages = [
+      'chewie',
+      'wakelock',
+    ];
+
+    // List top pub.dev packages
+    logger.info('Collecting top $count pub.dev packages...');
+    final topPubPackages = <String, String>{};
+    var page = 1;
+    while (topPubPackages.length < count) {
+      final results = await _pubClient.search('', page: page++);
+      for (final packageName in results.packages.map((pkg) => pkg.package)) {
+        if (unofficiallyAbandonedPackages.contains(packageName)) {
+          continue;
+        }
+
+        // Get latest version
+        final packageInfo = await _pubClient.packageInfo(packageName);
+        topPubPackages[packageName] = packageInfo.latest.version;
+      }
+    }
+
+    // Create app with all Amplify Flutter dependencies
+    logger.info('Creating temporary app...');
+    final appDir =
+        Directory.systemTemp.createTempSync('amplify_constraints_verify_');
+    final createRes = await Process.start(
+      'flutter',
+      ['create', '--project-name=constraints_verify', '.'],
+      workingDirectory: appDir.path,
+      mode: ProcessStartMode.inheritStdio,
+    );
+    if (await createRes.exitCode != 0) {
+      throw Exception('flutter create failed');
+    }
+    const amplifyPackages = [
+      'amplify_flutter',
+      'amplify_analytics_pinpoint',
+      'amplify_api',
+      'amplify_auth_cognito',
+      'amplify_datastore:^1.1.0-supports.only.mobile.0',
+      'amplify_storage_s3',
+    ];
+    final addRes = await Process.run(
+      'flutter',
+      ['pub', 'add', ...amplifyPackages],
+      workingDirectory: appDir.path,
+      stdoutEncoding: utf8,
+      stderrEncoding: utf8,
+    );
+    if (addRes.exitCode != 0) {
+      throw Exception(
+        'Could not add Amplify packages: ${addRes.stderr}',
+      );
+    }
+
+    // Try adding each package
+    final failedPackages = <String>[];
+    for (final MapEntry(key: packageName, value: latestVersion)
+        in topPubPackages.entries) {
+      logger.info('Verifying "$packageName:$latestVersion"...');
+      final addRes = await Process.run(
+        'flutter',
+        ['pub', 'add', '$packageName:$latestVersion'],
+        workingDirectory: appDir.path,
+        stdoutEncoding: utf8,
+        stderrEncoding: utf8,
+      );
+      if (addRes.exitCode != 0) {
+        failedPackages.add('$packageName:$latestVersion');
+        logger.error(
+          'Failed to add "$packageName"',
+          '${addRes.stdout}\n${addRes.stderr}',
+        );
+      }
+      final removeRes = await Process.run(
+        'flutter',
+        ['pub', 'remove', packageName],
+        workingDirectory: appDir.path,
+        stdoutEncoding: utf8,
+        stderrEncoding: utf8,
+      );
+      if (removeRes.exitCode != 0) {
+        throw Exception('Error removing package: ${removeRes.stderr}');
+      }
+    }
+
+    if (failedPackages.isNotEmpty) {
+      logger
+        ..error('Failed to add the following packages:')
+        ..error(failedPackages.map((pkg) => '- $pkg').join('\n'));
+      exit(1);
+    }
+
+    logger.info('All packages succeeded.');
+  }
+
+  @override
+  void close() {
+    _pubClient.close();
+    super.close();
   }
 }