+++ packages/amplify_core/lib/src/types/query/query_predicate.dart
@@ -58,9 +58,15 @@ class QueryPredicateOperation extends QueryPredicate {
   @override
   bool evaluate(Model model) {
     final fieldName = getFieldName(field);
-    //ignore:implicit_dynamic_variable
-    final value = model.toJson()[fieldName];
-    return queryFieldOperator.evaluate(value);
+    // TODO(Jordan-Nelson): Remove try/catch at next major version bump
+    try {
+      final value = model.toMap()[fieldName];
+      return queryFieldOperator.evaluate(value);
+    } on UnimplementedError {
+      final value = model.toJson()[fieldName];
+      // ignore: deprecated_member_use_from_same_package
+      return queryFieldOperator.evaluateSerialized(value);
+    }
   }
 
   @override