+++ packages/auth/amplify_auth_cognito/example/integration_test/user_attributes_test.dart
@@ -2,14 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-import 'package:amplify_auth_cognito_example/amplifyconfiguration.dart';
+import 'package:amplify_auth_integration_test/amplify_auth_integration_test.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:amplify_integration_test/amplify_integration_test.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'utils/setup_utils.dart';
-import 'utils/test_utils.dart';
+import 'test_runner.dart';
 
 extension on List<AuthUserAttribute> {
   String? valueOf(AuthUserAttributeKey authUserAttributeKey) =>
@@ -19,7 +18,7 @@ extension on List<AuthUserAttribute> {
 }
 
 void main() {
-  initTests();
+  testRunner.setupTests();
 
   group('User Attributes', () {
     for (final environmentName in userPoolEnvironments) {
@@ -30,9 +29,9 @@ void main() {
         late String phoneNumber;
         late String name;
 
-        setUpAll(() async {
-          await configureAuth(
-            config: amplifyEnvironments[environmentName]!,
+        setUp(() async {
+          await testRunner.configure(
+            environmentName: environmentName,
           );
 
           username = generateUsername();
@@ -41,7 +40,7 @@ void main() {
           phoneNumber = generatePhoneNumber();
           name = generateNameAttribute();
 
-          final cognitoUsername = await adminCreateUser(
+          await adminCreateUser(
             username,
             password,
             autoConfirm: true,
@@ -61,7 +60,6 @@ void main() {
               )
             ],
           );
-          addTearDown(() => deleteUser(cognitoUsername));
 
           final res = await Amplify.Auth.signIn(
             username: username,