+++ packages/auth/amplify_auth_cognito/example/integration_test/version_upgrade_test.dart
@@ -6,19 +6,19 @@ import 'dart:io';
 import 'package:amplify_api/amplify_api.dart';
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_auth_cognito_example/amplifyconfiguration.dart';
+import 'package:amplify_auth_integration_test/amplify_auth_integration_test.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:amplify_integration_test/amplify_integration_test.dart';
 import 'package:amplify_native_legacy_wrapper/amplify_native_legacy_wrapper.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:integration_test/integration_test.dart';
 
-import 'utils/test_utils.dart';
+import 'test_runner.dart';
 import 'utils/validation_utils.dart';
 
 final usernameConfig = amplifyEnvironments['sign-in-with-username']!;
 
 void main() {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+  testRunner.setupTests();
 
   group(
     'should migrate a user session from the legacy plugin',
@@ -26,9 +26,9 @@ void main() {
     () {
       final legacyPlugin = AmplifyNativeLegacyWrapper();
 
-      late final String username;
-      late final String cognitoUsername;
-      late final String password;
+      late String username;
+      late String cognitoUsername;
+      late String password;
 
       setUp(() async {
         // configure Amplify and legacy plugin, and ensure no users are signed in.
@@ -80,19 +80,17 @@ Future<void> configureAmplify(String config) async {
   if (Amplify.isConfigured) {
     await Amplify.reset();
   }
-  await Amplify.addPlugins([AmplifyAPI(), AmplifyAuthCognito()]);
+  await Amplify.addPlugins([AmplifyAPI(), AmplifyAuthTestPlugin()]);
   await Amplify.configure(config);
 }
 
-Future<String> createUser(String username, String password) async {
-  final cognitoUserName = await adminCreateUser(
+Future<String> createUser(String username, String password) {
+  return adminCreateUser(
     username,
     password,
     autoConfirm: true,
     verifyAttributes: true,
   );
-  addTearDown(() => deleteUser(cognitoUserName));
-  return cognitoUserName;
 }
 
 /// Clears the Amplify vLatest credential store, including the "version" key