+++ packages/auth/amplify_auth_cognito_dart/lib/src/model/hosted_ui/oauth_parameters.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of amplify_auth_cognito.hostedui.oauth_parameters;
+part of 'oauth_parameters.dart';
 
 // **************************************************************************
 // BuiltValueGenerator