+++ packages/authenticator/amplify_authenticator/example/integration_test/test_runner.dart
@@ -1,0 +1,5 @@
+import 'package:amplify_auth_integration_test/amplify_auth_integration_test.dart';
+import 'package:amplify_authenticator_example/amplifyconfiguration.dart';
+
+/// The global test runner.
+const AuthTestRunner testRunner = AuthTestRunner(amplifyEnvironments);