+++ packages/authenticator/amplify_authenticator/lib/src/widgets/button.dart
@@ -105,33 +105,43 @@ class _AmplifyElevatedButtonState
   Widget build(BuildContext context) {
     final buttonResolver = stringResolver.buttons;
     final loadingIndicator = widget.loadingIndicator;
+    final onPressed =
+        state.isBusy ? null : () => widget.onPressed(context, state);
+    final child = state.isBusy && loadingIndicator != null
+        ? loadingIndicator
+        : Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              if (widget.leading != null) widget.leading!,
+              Flexible(
+                child: FittedBox(
+                  fit: BoxFit.scaleDown,
+                  child: Text(
+                    buttonResolver.resolve(
+                      context,
+                      widget.labelKey,
+                    ),
+                    textAlign: TextAlign.center,
+                  ),
+                ),
+              ),
+              if (widget.trailing != null) widget.trailing!,
+            ].spacedBy(const SizedBox(width: 10)),
+          );
+    final useMaterial3 = Theme.of(context).useMaterial3;
     return SizedBox(
       width: double.infinity,
-      child: ElevatedButton(
-        focusNode: focusNode,
-        onPressed: state.isBusy ? null : () => widget.onPressed(context, state),
-        child: state.isBusy && loadingIndicator != null
-            ? loadingIndicator
-            : Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  if (widget.leading != null) widget.leading!,
-                  Flexible(
-                    child: FittedBox(
-                      fit: BoxFit.scaleDown,
-                      child: Text(
-                        buttonResolver.resolve(
-                          context,
-                          widget.labelKey,
-                        ),
-                        textAlign: TextAlign.center,
-                      ),
-                    ),
-                  ),
-                  if (widget.trailing != null) widget.trailing!,
-                ].spacedBy(const SizedBox(width: 10)),
-              ),
-      ),
+      child: useMaterial3
+          ? FilledButton(
+              focusNode: focusNode,
+              onPressed: onPressed,
+              child: child,
+            )
+          : ElevatedButton(
+              focusNode: focusNode,
+              onPressed: onPressed,
+              child: child,
+            ),
     );
   }
 }