+++ packages/auth/amplify_auth_cognito_dart/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.10.0+1
+
+### Fixes
+- fix(auth): User attribute serialization
+
 ## 0.10.0
 
 ### Features