+++ packages/aws_signature_v4/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.0+1
+
+- Minor bug fixes and improvements
+
 ## 0.4.0
 
 ### Features